refactored... even... more.... ahhhhgit status
